Made QuartoGameTransitionGenerator class
